reorganize comms
