+++ src/dx/__init__.py
@@ -1,3 +1,5 @@
+from IPython import get_ipython
+
 from .comms import *
 from .dx import *
 from .formatters import *
@@ -9,3 +11,8 @@ __version__ = "1.2.0"
 
 configure_logging()
 set_display_mode("simple")
+
+if (ipython := get_ipython()) is not None:
+    register_resampler_comm(ipython)
+    register_renamer_comm(ipython)
+    register_exporter_comm(ipython)
